requirements.txt
streamlit
streamlit-option-menu
pandas
openpyxl
qrcode
pillow
matplotlib
zxing-cpp
